==> Data processing/data_cleaning/1_data_combine.ipynb <==
import json
import pandas as pd

pd_all = None

with open("../json_files/story_data.json", "r") as file:
    data = json.loads(file.read())
    pd_all = pd.json_normalize(data)
    
len(pd_all)

for i in range(1, 4):
    with open("../json_files/depth_{}.json".format(i), "r") as file:
        data = json.loads(file.read())
        pd_new = pd.json_normalize(data)
        pd_all = pd.concat([pd_new, pd_all])

len(pd_all)

pd_all.to_csv("./1_data_combine.csv", index=False)
