==> Data processing/data_cleaning/4_combine_description_and_features.ipynb <==
import pandas as pd

df_description = pd.read_csv("./3_crawl_description.csv")

df_features = pd.read_csv("./2_data_cleaning.csv")

df_description.head()

df_features.head()

df_features_final = pd.merge(df_description, df_features, on="ID").drop(["link"], axis=1)

df_features_final.to_csv("./4_combine_description_and_features.csv", index=False)

df_features_final["depth"].value_counts()

len(df_features_final)
