==> Data processing/data_cleaning/5_crawl_tags_and_clean_null_description.ipynb <==
import pandas as pd
import numpy as np
import math

df_features = pd.read_csv("./4_combine_description_and_features.csv")

df_id = pd.read_csv("./2_id_with_link.csv")

# Loc null description & full tags

blank_id = []
crawl_id = []

for i in range(len(df_features)):
    tags = df_features.at[i, "story_tags_list"]
    description = df_features.at[i, "Description"]
    id_story = df_features.at[i, "ID"]
    
    if pd.isnull(description):
        blank_id.append(id_story)
        continue
    
    if type(tags) == type(np.nan) or type(tags) == type(math.nan):
        crawl_id.append(id_story)
        continue
    
    if len(tags) < 5:
        crawl_id.append(id_story)
        continue

print(blank_id)
print(crawl_id)

df_loc_description = df_features.loc[~df_features["ID"].isin(blank_id)]

df_to_crawl_tags = df_loc_description.loc[df_loc_description["ID"].isin(crawl_id)]

df_to_crawl_tags_with_link = df_id.loc[df_id["ID"].isin(df_to_crawl_tags["ID"])]

df_to_crawl_tags_with_link["tags"] = np.nan

df_to_crawl_tags_with_link = df_to_crawl_tags_with_link.reset_index().drop(["index"], axis=1)

df_to_crawl_tags_with_link

def login(driver, timeout_value=10000):
    # Access the variables using os.environ.get
    USERNAME = ""
    PASSWORD = ""

    login_button = WebDriverWait(driver, timeout_value).until(
        EC.presence_of_element_located((By.XPATH, "/html/body/div[3]/div/div[3]/div/div[2]/main/div/div/div/div/button"))
    )

    # Click the button
    login_button.click()

    username_fill = WebDriverWait(driver, timeout_value).until(
        EC.presence_of_element_located((By.XPATH, "/html/body/div[3]/div/div[3]/div/div[2]/main/div/div/div/div/form/div/div[1]/input"))
    )

    username_fill.send_keys(USERNAME)

    password_fill = WebDriverWait(driver, timeout_value).until(
        EC.presence_of_element_located((By.XPATH, "/html/body/div[3]/div/div[3]/div/div[2]/main/div/div/div/div/form/div/div[2]/div/input"))
    )

    password_fill.send_keys(PASSWORD)

    login_button = WebDriverWait(driver, timeout_value).until(
        EC.presence_of_element_located((By.XPATH, "/html/body/div[3]/div/div[3]/div/div[2]/main/div/div/div/div/form/input"))
    )

    login_button.click()
    
    return driver

from selenium import webdriver
from selenium.webdriver.common.by import By
from selenium.webdriver.support.ui import WebDriverWait
from selenium.webdriver.support import expected_conditions as EC
from selenium.webdriver.chrome.options import Options

import os
import time

#from dotenv import load_dotenv

#from functions import login, nav_tags, close_db, clear_db, reopen_db, stories_crawled, old_stories

###### Environment variables ######
# Load environment variables from the .env file
#load_dotenv()

###### Setup chrome driver ######
# Ignore certificate errors
options = Options()
options.add_argument('--ignore-certificate-errors')
options.add_experimental_option('excludeSwitches', ['enable-logging'])

# Set the path to your chromedriver executable
chrome_driver_path = os.path.join(os.getcwd(), "other", "chromedriver-win64", "chromedriver.exe")

# Create a new instance of the Chrome driver
driver = webdriver.Chrome(options=options)

# Open the webpage
driver.get("https://www.wattpad.com/login")

timeout_value = 10000

driver = login(driver, timeout_value)

for i in range(len(df_to_crawl_tags_with_link)):
    if pd.isna(df_to_crawl_tags_with_link.iloc[i]["tags"]) == False:
        print(i)
        continue 
        
    link = df_to_crawl_tags_with_link.iloc[i]["link"]
    index = df_to_crawl_tags_with_link.iloc[i]["ID"]
    
    driver.get(link)
    
    try:
        tags_wrapper = WebDriverWait(driver, 10).until(
            EC.presence_of_element_located((By.CLASS_NAME, "tag-items"))
        )
        
        each_tags = WebDriverWait(tags_wrapper, 10).until(
            EC.presence_of_all_elements_located((By.CLASS_NAME, "tag-pill__yrFxu"))
        )
        
        extract_tags = [each_tag.text for each_tag in each_tags]
        
        print(str(extract_tags))
        
        #break
    
        df_to_crawl_tags_with_link.at[i, "tags"] = str(extract_tags)

        if i % 5 == 4:
            print(f"Crawl to index {i}")
    except Exception as e:
        print(e)
    except:
        print(f"Error index {i}")
        id_error.append(i)

len(df_to_crawl_tags_with_link.loc[df_to_crawl_tags_with_link["tags"].isnull()])

for i in range(len(df_to_crawl_tags_with_link)):
    id_story = df_to_crawl_tags_with_link.at[i, "ID"]
    tags_story = df_to_crawl_tags_with_link.at[i, "tags"]
    
    #print(id_story)
    #print(tags_story)
    
    id_df = df_loc_description.loc[df_loc_description["ID"] == id_story].index[0]
    
    df_loc_description.at[id_df, "story_tags_list"] = tags_story

len(df_loc_description.loc[df_loc_description["story_tags_list"].isnull()])

df_loc_description = df_loc_description.loc[df_loc_description["story_tags_list"].isnull() == False]
df_loc_description

df_loc_description.to_csv("./5_crawl_tags_and_clean_null_description.csv", index=False)
