==> Data processing/data_cleaning/6_filter_latin_tags.ipynb <==
import pandas as pd
import re
import numpy as np
from collections import Counter
import unicodedata

df = pd.read_csv("./5_crawl_tags_and_clean_null_description.csv")

df.head()

regex = re.compile('[^\u0000-\u007F\u0080-\u00FF\u0100-\u024F\u1E00-\u1EFF\u0300-\u036F]')

def clean_up(text):
    if type(text) == type(np.nan):
        return text
    
    # Clean out non-Latin characters
    text = regex.sub('', text)
    
    # Clean out - symbol
    text = re.sub(r'-', '', text)
    
    return text

df['story_tags_list'] = df['story_tags_list'].apply(clean_up)

def remove_vietnamese_diacritics(text):
    if type(text) == type(np.nan):
        return text
    
    text = text.replace('đ', 'd')
    # Normalize the text to decomposed Unicode form
    normalized_text = unicodedata.normalize('NFD', text)
    # Use regex to remove combining diacritical marks
    cleaned_text = re.sub('[\u0300-\u036f]', '', normalized_text)
    return cleaned_text


df['story_tags_list'] = df['story_tags_list'].apply(remove_vietnamese_diacritics)

def str_to_list(string):
    if isinstance(string, list):
        return string
    return string[2:-2].split("', '")

tags_list = []

for i in range(len(df)):
    tags = df.iloc[i]["story_tags_list"]
        
    new_tags = []
    delete_tags = False
    
    if type(tags) == type(np.nan):
        print(tags)
        continue
        
    df.at[i, "story_tags_list"] = str_to_list(tags)
    
    for tag in str_to_list(tags):
        if len(tag) == 0:
            delete_tags = True
            continue
        
        new_tags.append(tag)
    
    if delete_tags:
        df.at[i, "story_tags_list"] = new_tags
    
    tags_list.extend(new_tags)

Counter(tags_list)

df.to_csv("./6_filter_latin_tags.csv", index=False)

len(tags_list)

len(Counter(tags_list).keys())

for i in range(len(df)):
    tags = df.at[i, "story_tags_list"]
    
    if type(tags) == type(np.nan):
        print(f"0: {i}")
        continue
        
    #tags = str_to_list(tags)
        
    if len(tags) == 0:
        print(f"1: {i}")
        print(tags)
        continue
        
    if len(tags) == 1 and len(tags[0]) == 1:
        print(f"2: {i}")
        continue

deleted = [398, 1560]

df = df.loc[~df.index.isin(deleted)]

df.to_csv("./6_filter_latin_tags.csv", index=False)
