==> Data processing/data_cleaning/8_filter_latin_description.ipynb <==
import pandas as pd
import re
import numpy as np
from collections import Counter
import unicodedata

df = pd.read_csv("./7_tags_grouping.csv")

df.head()

regex = re.compile('[^\u0000-\u007F\u0080-\u00FF\u0100-\u024F\u1E00-\u1EFF\u0300-\u036F]')

def clean_up(text):
    if type(text) == type(np.nan):
        return text
    
    # Clean out non-Latin characters
    text = regex.sub('', text)
    
    return text

df['Description'] = df['Description'].apply(clean_up)

for i in range(len(df)):
    #print(len(df.at[i, "Description"]))
    
    string_to_del = "Bảo Lưu Mọi Quyền"
    
    description = df.iloc[i]["Description"]
    
    if type(description) == type(np.nan):
        continue
    
    if description[-len(string_to_del):] == "Bảo Lưu Mọi Quyền":
        print(i)
        df.at[i, "Description"] = description[:-len(string_to_del)]

for i in range(len(df)):
    #print(len(df.at[i, "Description"]))
    description = df.iloc[i]["Description"]
    
    if type(description) == type(np.nan):
        continue
    
    if len(description) < 10:
        print(i, len(description))
        print(description)
        
    if len(description) == 0:
        df.at[i, "Description"] = np.nan

df = df.loc[df["Description"].isnull() == False]

#df.loc[df["story_tags_list"].str.contains()]

df.to_csv("./8_filter_latin_description.csv", index=False)

df
