==> Data processing/data_cleaning/9_deleted_without_latin_description.ipynb <==
import pandas as pd
import re
import numpy as np
from collections import Counter
import unicodedata

df = pd.read_csv("./8_filter_latin_description.csv")

id_to_remove = [121,  398,  400,  487,  700,  757,  758,  769,  785,  851,  866,
       1007, 1009, 1029, 1095, 1149, 1205, 1216, 1222, 1238, 1295, 1300,
       1445, 1446, 1515, 1540, 1577, 1578, 1687, 2109, 2354, 2432]

df = df.loc[~df["ID"].isin(id_to_remove)]

df.to_csv("./9_deleted_without_latin_description.csv", index=False)
