==> Data processing/data_cleaning/10_merge_annotated_tags.ipynb <==
import pandas as pd

df_orig = pd.read_csv("./9_deleted_without_latin_description.csv")

df_to_merge = pd.read_excel("story_annotated.xlsx")

df_to_merge[["ID", "story_tags_list"]]

for i in range(len(df_to_merge)):
    id_story = df_to_merge.at[i, "ID"]
    
    index_df_orig = df_orig.loc[df_orig["ID"] == id_story]
    
    if index_df_orig.empty:
        continue
    
    index_df_orig_index = index_df_orig.index[0]
    
    
    df_orig.at[index_df_orig_index, "story_tags_list"] = df_to_merge.at[i, "story_tags_list"]

df_orig

df_orig.to_csv("./10_merge_annotated_tags.csv", index=False)
