==> Data processing/data_cleaning/11_tags_grouping_after_annotated.ipynb <==
import pandas as pd
from collections import Counter

df = pd.read_csv("./10_merge_annotated_tags.csv")

tags_hash_table = {
    1: ["romance", "love", "langman", "ngot", "sung", "ngontinh", "tinhcam", "lovestory", "teenromance", "loveatfirstsight", 
        "chicklit", "heartbreak", "yeu", "thoca"],
    2: ["drama", "kinhdi", "horror", "nguoisoi", "macarong", "wolves", 'werewolf', 'paranormal', 'werewolves', 'pack', 
        "adventure", "phieuluu", "bian", "mystery", "secrets", "mysterious", "action", "hanhdong", "thriller", "trinhtham", 
        "killer", "phaan", "doan"],
    3: ["fanfic", "fanfiction", "fandom", "teenfic", "teenfiction", "fiction", "fairytail", "fic", "hucau", 
        "fantasy", "vientuong", "supernatural", "magic", "truyentamlinh", "sieunhien"],
    4: ["lgbt", "lesbian", "dammy", "bl", "bachhop", "boyxboy", "gay", "yaoi", "boylove", "girllove", "freethelgbt", "bl", 
        "gl", "songtinh", "kpop", "bts", 'taehyung', 'vkook', 'jungkook', "dammy"],
    5: ["haihuoc", "humor", "comedy", "funny"],
    6: ["hiendai", "modern", "school", "hocduong", "highschool", "teen", "youngadult"],
    7: ["codai", "classics"],
    8: ["tieuthuyet", "tieuthuyetlichsu", "tieuthuyetthieunien", "novel", "tieuthuyetchung"],
}

tags_hash_key = {}

# Function to check if a string contains lowercase letters
def contains_lowercase(text):
    return any(c.islower() for c in text)

# Apply the function to the DataFrame column
df_only_tags = df.loc[df['story_tags_list'].apply(contains_lowercase)].reset_index().drop(["index"], axis=1)

len(df_only_tags)

df_only_tags

# key value swap
def key_value_swap():
    for key, value in tags_hash_table.items():
        for tag in value:
            tags_hash_key[tag] = key

# tags: str to list
def str_to_list(string):
    if len(string) <= 4:
        return list()
    return string[2:-2].split("', '")

def tags_str_to_list():
    for i in range(len(df_only_tags)):
        tags = df_only_tags.iloc[i]["story_tags_list"]

        if isinstance(tags, list):
            continue

        df_only_tags.at[i, "story_tags_list"] = str_to_list(tags)

def tags_string_to_num(printing=True):
    full_num = 0
    
    key_value_swap()
    tags_str_to_list()

    for i in range(len(df_only_tags)):
        tags = df_only_tags.iloc[i]["story_tags_list"]
        
        full_num_bool = False

        for index, tag in enumerate(tags):
            if isinstance(tag, int):
                full_num_bool = True
                break

            if tag in tags_hash_key.keys():
                full_num_bool = True
                break
        
        if full_num_bool:
            new_tags_list = []
            for index, tag in enumerate(tags):
                if isinstance(tag, int):
                    new_tags_list.append(tag)
                    continue
              
                if tag in tags_hash_key.keys():
                    new_tags_list.append(tags_hash_key[tag])
                    continue
                
            full_num += 1       
                
            df_only_tags.at[i, "story_tags_list"] = new_tags_list

    print(full_num)
    print(len(df))
    
tags_string_to_num()

def all_unique_tags():
    #tags_string_to_num(printing=False)
    
    unique_tags = []
    
    for i in range(len(df_only_tags)):
        tags = df_only_tags.iloc[i]["story_tags_list"]
        
        unique_tags.extend(tags)
        
    #print(len(Counter(unique_tags).keys()))
    #print(Counter(unique_tags))
    
    tags_with_freq = {key: value for key, value in Counter(unique_tags).items() if isinstance(key, str) and value >= 50}
    
    print(len(Counter(unique_tags)))
    print(Counter(unique_tags))
    print(len(tags_with_freq))
    print(tags_with_freq)

all_unique_tags()

for i in range(len(df_only_tags)):
    id_story = df_only_tags.at[i, "ID"]
    
    index_df_orig = df.loc[df["ID"] == id_story]
    
    if index_df_orig.empty:
        continue
    
    index_df_orig_index = index_df_orig.index[0]
    
    
    df.at[index_df_orig_index, "story_tags_list"] = df_only_tags.at[i, "story_tags_list"]

for i in range(len(df)):
    tags = df.iloc[i]["story_tags_list"]
    
    if isinstance(tags, str):
        continue
        
    if pd.isna(tags).any() == True:
        continue
        
    df.at[i, "story_tags_list"] = str(tags)

df

df.to_csv("./11_tags_grouping_after_annotated_part_1.csv", index=False)
df_only_tags.to_csv("./11_tags_to_annotated.csv", index=False)
