==> Data processing/data_cleaning/12_group_last_time.ipynb <==
import pandas as pd

df_orig = pd.read_csv("./11_tags_grouping_after_annotated_part_1.csv")

df_to_combine = pd.read_excel("./11_tags_have_been_annotated.xlsx")

df_to_combine

for i in range(len(df_to_combine)):
    id_story = df_to_combine.at[i, "ID"]
    
    index_df_orig = df_orig.loc[df_orig["ID"] == id_story]
    
    if index_df_orig.empty:
        print("here")
        continue
    
    index_df_orig_index = index_df_orig.index[0]
    
    
    df_orig.at[index_df_orig_index, "story_tags_list"] = df_to_combine.at[i, "story_tags_list"]

df_orig

df_orig.to_csv("./12_group_last_time.csv", index=False)

df_orig.loc[df_orig["ID"] == 455, "story_tags_list"]
