==> Data processing/data_cleaning/14_delete_duplicate_group.ipynb <==
import pandas as pd

df = pd.read_csv("./13_tags_grouping_part_2.csv")

def str_to_list(string):
    new_list = string[1:-1].split(', ')
    
    new_list_final = []
    
    for item in new_list:
        conv = 0
        
        try:
            conv = int(item)
            new_list_final.append(conv)
        except:
            pass
        
    new_list_final = list(set(new_list_final))    
        
    return str(new_list_final)

df["story_tags_list"] = df["story_tags_list"].apply(str_to_list)

df.to_csv("./14_delete_duplicate_group.csv", index=False)
