==> Data processing/data_cleaning/15_clean_link_inside_description.ipynb <==
import pandas as pd
import numpy as np
import re

df = pd.read_csv("./14_delete_duplicate_group.csv")

df

pattern = r'\b(?:https?://)?(?:www\.)?[\w-]+\.[a-z]{2,}(?:\.[a-z]{2})?\b'
pattern_2 = r'[-a-zA-Z0-9@:%._\+~#=]{1,256}\.[a-zA-Z0-9()]{1,6}\b([-a-zA-Z0-9()@:%_\+.~#?&//=]*)'

df.loc[df["Description"].str.contains(pattern_2)]

df.loc[df["Description"].str.contains(pattern)]

def remove_links(text):
    pattern_2 = r'[-a-zA-Z0-9@:%._\+~#=]{1,256}\.[a-zA-Z0-9()]{1,6}\b([-a-zA-Z0-9()@:%_\+.~#?&//=]*)'
    cleaned_text = re.sub(pattern_2, '', text)
    return cleaned_text

df["Description"] = df["Description"].apply(remove_links)

df.to_csv("./15_clean_link_inside_description.csv")
