==> Data processing/data_cleaning/17_one_hot_class.ipynb <==
import numpy as np
import pandas as pd
import random

df = pd.read_csv("./16_description_cleaned.csv")

df.at[0, "tag"]

def str_to_list(string):
    try:
        return_arr = [int(i) for i in string[1:-1].split(" ")]
    except Exception as e:
        return list()
    return return_arr

df["tag"] = df["tag"].apply(str_to_list)

def list_to_class(lst_class):
    final_arr = []
    for i in range(1, 9):
        final_arr.append(i in lst_class)
    
    return final_arr

df["tag"] = df["tag"].apply(list_to_class)

df["predicted_tag"] = "None"

for i in range(1, 9):
    df[str(i)] = np.nan

for i in range(len(df)):
    for j in range(1, 9):
        df.at[i, str(j)] = df.at[i, "tag"][j - 1]

df = df.drop(["tag", "predicted_tag"], axis=1)

df.to_csv("./17_one_hot_class.csv", index=False)
